IMPALA-2905: Handle coordinator fragment lifecycle like all others
The plan-root fragment instance that runs on the coordinator should be
handled like all others: started via RPC and run asynchronously. Without
this, the fragment requires special-case code throughout the
coordinator, and does not show up in system metrics etc.

This patch adds a new sink type, PushPullSink, to the root fragment
instance so that the coordinator can pull row batches that are pushed by
the root instance. The coordinator signals completion to the fragment
instance via closing the consumer side of the sink, whereupon the
instance is free to complete.

Since the root instance now runs asynchronously wrt to the coordinator,
we add several coordination methods to allow the coordinator to wait for
a point in the instance's execution to be hit - e.g. to wait until the
instance has been opened.

Done in this patch:

* Add PushPullSink
* Add coordination to PFE to allow coordinator to observe lifecycle
* Allow movable types to be used in BlockingQueue
* Make FragmentMgr a singleton
* Removed dead code from Coordinator::Wait()
* Decoupled QES output exprs from PFE's LLVM module by introducing
  temporary RuntimeState.
* Remove special-case limit-based teardown of coordinator fragment, and
  supporting functions in PlanFragmentExecutor.

Not yet done:

* Fix planner tests to reflect new sink added at root.

Some effort has been made to limit the impact on Coordinator to
alleviate the effect of conflicts.

TODO in future patches:

* Move output expr evaluation into PushPullSink
* Clean up fragment profile initialisation in coordinator.

Testing:

EE tests fail predictably in test_tpcds_queries (although memtracker
tests may be implicated). Not yet reproducible locally. Patch submitted
for review to address core logic changes.

Change-Id: Ibb0064ec2f085fa3a5598ea80894fb489a01e4df