IMPALA-5001: Redownload dependencies if toolchain ID changes
To allow for the situation where a dependency's compile flags
change (but otherwise nothing else does, including the version), this
patch adds the toolchain build ID to the version information checked by
bootstrap_toolchain.py when it decides whether to download a new
dependency version.

Also add a small amount of logging to make it clearer what the script is
doing.

Change-Id: I77fcfbeb4594b14da2048a9c31458f634600c1e6