[SECURITY] Enable MiniKdc and add kerberized BE tests
This patch enables the MiniKdc and converts the rpc-mgr-test
to have a kerberized mode, which uses the MiniKdc to run the tests
configured with Kerberos enabled.

There are a couple of small code changes on the KuduRPC side as
those changes are currently required to avoid a clash between
Impala's and Kudu's util libraries.