IMPALA-5592: Fix computation of uncompressed row batch size
The uncompressed row batch size is the size of a row batch
if the tuple data is sent without compression. It's computed
by the following formula:

serialized row batch size - compressed tuple data size + uncompressed tuple data size

The existing code is using the wrong field for the compressed tuple
data size and it causes uncompressed data size to be always identical
to the compressed row batch size. This change fixes the problem by
using the right field for the compressed tuple data size.

Change-Id: Ic8d285d1e4a1aa97c98c505351c18fea04c60311